fix: perform explicit typecast to avoid sign conversion warnings
* Fix sign-conversion warnings by making casts/constant types
  explicit in FF-A call arguments, masks, and error paths.

* Additionally address static-check findings triggered by these
  changes, including uninitialized variables, pointer arithmetic
  flagged as undefined behavior in tests and dead stores.

* Some large TEST_PRECONDITION() expansions trigger the
  readability-function-size clang-tidy check. These tests are
  intentionally large integration-style tests, so the warning is
  suppressed locally with NOLINTNEXTLINE.

+++ inc/vmapi/hf/call.h
@@ -658,7 +658,7 @@ static inline struct ffa_value ffa_notification_get(ffa_id_t receiver_vm_id,
 {
 	return ffa_call((struct ffa_value){
 		.func = FFA_NOTIFICATION_GET_32,
-		.arg1 = (vcpu_id << 16) | (receiver_vm_id),
+		.arg1 = ((uint64_t)vcpu_id << 16) | (receiver_vm_id),
 		.arg2 = flags,
 	});
 }
